Removing Rounds Concept
The idea of adding additional "rounds" does not have the desired affect
on cost for streaming solve. This is because a clever implementer could
make their PRNG forkable and instead of having to stream through all
rounds again, they can just stream through the last round negating the
cost of the extra rounds during the solve phase.